Make "move safety" try to continue in last direction.
The vote "had participation" if at least one person voted for any move.

If the vote had participation, apply the top voted move.

If the vote did not have any participation, the "top voted move" is
either none (first turn's vote) or the last moved direction.

If the vote did not have any participation, prefer moving in the last
moved direction as long as there is one and as long as it's considered
safe, otherwise pick a random safe move.